Rename browser session telemetry logs
